corrections and better flux

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: fde-capu <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/23 00:25:24 by fde-capu          #+#    #+#             */
-/* U20200214133458 ||::::                      */
+/* U20200215172727 ::|||:                      */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,11 +33,11 @@ static int	st_sum_dign(char *a)
 
 long long	ft_atoi(const char *str)
 {
-	char	*a;
-	char	*b;
-	int		l;
-	long	r;
-	int		neg;
+	char		*a;
+	char		*b;
+	int			l;
+	long long	r;
+	int			neg;
 
 	a = ft_strtrim((char *)str, " \t\n\r\v\f");
 	neg = *a == '-' ? 1 : 0;
@@ -56,5 +56,5 @@ long long	ft_atoi(const char *str)
 		a++;
 	}
 	r *= neg ? -1 : 1;
-	return ((int)r);
+	return (r);
 }
